+++ packages/remirror__extension-react-tables/src/table-extensions.tsx
@@ -167,7 +167,7 @@ export class TableExtension extends BaseTableExtension {
         table,
       });
 
-      replaceNodeAtPosition({ pos, tr, content: controlledTable });
+      replaceNodeAtPosition({ pos: tr.mapping.map(pos), tr, content: controlledTable });
     }
 
     dispatch(tr.setMeta('addToHistory', false));
@@ -264,7 +264,7 @@ export class TableExtension extends BaseTableExtension {
             table,
           });
 
-          replaceNodeAtPosition({ pos, tr, content: controlledTable });
+          replaceNodeAtPosition({ pos: tr.mapping.map(pos), tr, content: controlledTable });
         }
 
         return tr.steps.length > 0 ? tr : undefined;