+++ packages/remirror__react-editors/src/types.ts
@@ -1,8 +1,9 @@
 import type { CreateEditorStateProps } from 'remirror';
-import type { RemirrorProps } from '@remirror/react';
+import type { RemirrorProps, UseThemeProps } from '@remirror/react';
 
 export interface ReactEditorProps
   extends Pick<CreateEditorStateProps, 'stringHandler'>,
     Pick<RemirrorProps, 'initialContent' | 'editable' | 'autoFocus' | 'hooks'> {
   placeholder?: string;
+  theme?: UseThemeProps['theme'];
 }