+++ packages/remirror__types/src/index.ts
@@ -1,4 +1,15 @@
+import type { Writable } from 'type-fest';
+
 export * from './types';
+
+/** @deprecated Use built-in `Awaited` instead */
+type PromiseValue<T> = Awaited<T>;
+
+/** @deprecated Renamed to `Writable` */
+type Mutable<BaseType, Keys extends keyof BaseType = keyof BaseType> = Writable<BaseType, Keys>;
+
+export type { Mutable, PromiseValue };
+
 export type {
   Asyncify,
   AsyncReturnType,
@@ -21,7 +32,6 @@ export type {
   LiteralUnion,
   Merge,
   MergeExclusive,
-  Mutable,
   ObservableLike,
   Opaque,
   PackageJson,
@@ -29,7 +39,6 @@ export type {
   PascalCase,
   Primitive,
   Promisable,
-  PromiseValue,
   ReadonlyDeep,
   RequireAtLeastOne,
   RequireExactlyOne,
@@ -42,4 +51,5 @@ export type {
   TypedArray,
   UnionToIntersection,
   ValueOf,
+  Writable,
 } from 'type-fest';