+++ packages/remirror__react-editors/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @remirror/react-editors
 
+## 1.0.34
+
+> 2023-05-21
+
+### Patch Changes
+
+- Updated dependencies [e59e75aec]
+- Updated dependencies [b12e3fbfe]
+  - remirror@2.0.32
+
+## 1.0.33
+
+> 2023-05-01
+
+### Patch Changes
+
+- a690b4536: Add a new prop `theme` to `MarkdownEditor`, `SocialEditor` and `WysiwygEditor`.
+
 ## 1.0.32
 
 > 2023-04-27